Move IdentityFnTrait trait to primitives.
