+++ pallets/common/src/benchs/user_builder.rs
@@ -14,14 +14,14 @@
 // along with this program. If not, see <http://www.gnu.org/licenses/>.
 use crate::{
     benchs::{SecretKey, User},
-    traits::identity::{Config, IdentityFnTrait},
+    traits::identity::Config,
 };
 use schnorrkel::{ExpansionMode, MiniSecretKey};
 
 use codec::{Decode, Encode};
 use frame_support::traits::Currency;
 use frame_system::RawOrigin;
-use polymesh_primitives::traits::group::GroupTrait;
+use polymesh_primitives::traits::{group::GroupTrait, IdentityFnTrait};
 use polymesh_primitives::IdentityId;
 use sp_io::hashing::blake2_256;
 use sp_std::prelude::*;
